pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halluciGuard"
version = "0.1.0"
authors = [
    {name = "HalluciGuard Contributors"},
]
description = "Open-source AI hallucination detection middleware for LLM pipelines"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "AGPL-3.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "llm", "hallucination", "safety", "reliability", "openai", "anthropic"]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.scripts]
halluciGuard = "halluciGuard.cli_module:cli_entry"

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.25.0"]
google = ["google-genai>=1.0.0"]
server = ["fastapi>=0.100.0", "uvicorn>=0.20.0"]
langchain = ["langchain-core>=0.1.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-mock",
    "black",
    "ruff",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "google-genai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "langchain-core>=0.1.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "google-genai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "langchain-core>=0.1.0",
]

[tool.setuptools]
packages = ["halluciGuard"]
